code-model/requirements.txt
Flask==2.3.3
opencv-python==4.8.1.78
mediapipe==0.10.7
tensorflow==2.13.0
numpy==1.24.3
Pillow==10.0.1
Werkzeug==2.3.7 